+++ changelog.md
@@ -6,6 +6,7 @@
 <summary>Fixes and minor changes</summary>
 
 - The Valley staircase algorithm has been rewritten and optimized;
+- Added displaying of the color names in the preset editor; 
 - Fixed reading air and unexisting blocks from complete_palette.json;
 - Fixed mouse scroll behavior in the preset editor above and below the window box;